Backend/requirements.txt
numpy
pandas
ipykernel
requests
seaborn
joblib
matplotlib
flask
flask-cors

python-dotenv

# LangChain + Google Gemini
langchain
langchain-community
langchain-google-genai
google-generativeai

# PDF Extraction (Using PyPDFLoader)
pypdf

# Embeddings (Free Open-Source Model)
sentence-transformers

# Vector Database (ChromaDB - Local)
chromadb
langchain-groq